Refactor ShowFanArt.py -> fan_art.py
